refactor: move message handling to separate module
